# eslint + prettier caching & speed improvements