--- a/packages/app-mobile/package.json
+++ b/packages/app-mobile/package.json
@@ -9,7 +9,7 @@
     "web": "expo start --web",
     "add": "expo add",
     "eject": "expo eject",
-    "lint": "eslint ./src --ext .js,.jsx,.ts,.tsx",
+    "lint": "eslint ./src --ext .js,.jsx,.ts,.tsx --cache",
     "lint:fix": "yarn run lint --fix",
     "postinstall": "npx patch-package; cd ../../ && patch-package"
   },