--- a/packages/background/package.json
+++ b/packages/background/package.json
@@ -9,8 +9,8 @@
     "build": "tsc",
     "build:html": "parcel build src/service-worker-loader.html --dist-dir build --no-source-maps",
     "dev": "tsc --watch",
-    "lint": "eslint ./src --ext .js,.jsx,.ts,.tsx",
-    "lint:fix": "yarn run lint --fix",
+    "lint": "eslint ./src --ext .js,.jsx,.ts,.tsx --cache",
+    "lint:fix": "yarn run lint --fix --cache",
     "start": "parcel src/service-worker-loader.html -p 9333"
   },
   "dependencies": {