--- a/packages/background/package.json
+++ b/packages/background/package.json
@@ -10,7 +10,7 @@
     "build:html": "parcel build src/service-worker-loader.html --dist-dir build --no-source-maps",
     "dev": "tsc --watch",
     "lint": "eslint ./src --ext .js,.jsx,.ts,.tsx --cache",
-    "lint:fix": "yarn run lint --fix --cache",
+    "lint:fix": "yarn run lint --fix",
     "start": "parcel src/service-worker-loader.html -p 9333"
   },
   "dependencies": {