--- a/packages/common-public/.eslintrc.js
+++ b/packages/common-public/.eslintrc.js
@@ -1,4 +1,7 @@
 module.exports = {
   root: true,
   extends: ["custom"],
+  parserOptions: {
+    ecmaVersion: 6,
+  },
 };