--- a/packages/common-public/package.json
+++ b/packages/common-public/package.json
@@ -12,8 +12,8 @@
   "scripts": {
     "build": "./scripts/config.sh && tsc && tsc-alias && tsc -p tsconfig.rn.json",
     "dev": "tsc --watch",
-    "lint": "eslint ./src --ext .js,.jsx,.ts,.tsx",
-    "lint:fix": "yarn run lint --fix",
+    "lint": "eslint ./src --ext .js,.jsx,.ts,.tsx --cache",
+    "lint:fix": "yarn run lint --fix --cache",
     "start": "yarn build"
   },
   "dependencies": {