--- a/packages/common-public/package.json
+++ b/packages/common-public/package.json
@@ -13,7 +13,7 @@
     "build": "./scripts/config.sh && tsc && tsc-alias && tsc -p tsconfig.rn.json",
     "dev": "tsc --watch",
     "lint": "eslint ./src --ext .js,.jsx,.ts,.tsx --cache",
-    "lint:fix": "yarn run lint --fix --cache",
+    "lint:fix": "yarn run lint --fix",
     "start": "yarn build"
   },
   "dependencies": {