--- a/packages/common-public/src/index.ts
+++ b/packages/common-public/src/index.ts
@@ -2,8 +2,8 @@
 // publicly. We should remove this eventually.
 
 export * from "./constants";
-export * from "./types";
 export * from "./logging";
+export * from "./types";
 export * from "./utils";
 export * from "./zustand-store";