--- a/packages/react-xnft-dom-renderer/package.json
+++ b/packages/react-xnft-dom-renderer/package.json
@@ -11,7 +11,7 @@
   ],
   "scripts": {
     "build": "parcel build src/index.tsx --dist-dir dist",
-    "lint": "eslint ./src --ext .js,.jsx,.ts,.tsx",
+    "lint": "eslint ./src --ext .js,.jsx,.ts,.tsx --cache",
     "lint:fix": "yarn run lint --fix"
   },
   "dependencies": {