--- a/packages/wallet-standard/.eslintignore
+++ b/packages/wallet-standard/.eslintignore
@@ -1,1 +1,2 @@
 lib
+node_modules