--- a/packages/wallet-standard/.eslintrc
+++ b/packages/wallet-standard/.eslintrc
@@ -1,16 +1,8 @@
 {
     "root": true,
-    "extends": [
-        "eslint:recommended",
-        "plugin:@typescript-eslint/recommended",
-        "plugin:prettier/recommended"
-    ],
+    "extends": ["eslint:recommended", "plugin:@typescript-eslint/recommended", "prettier"],
     "parser": "@typescript-eslint/parser",
-    "plugins": [
-        "@typescript-eslint",
-        "prettier",
-        "require-extensions"
-    ],
+    "plugins": ["@typescript-eslint", "require-extensions"],
     "rules": {
         "@typescript-eslint/consistent-type-imports": "error"
     }