--- a/packages/wallet-standard/package.json
+++ b/packages/wallet-standard/package.json
@@ -16,7 +16,7 @@
         "clean": "shx mkdir -p lib && shx rm -rf lib",
         "tsc": "tsc --build --verbose tsconfig.root.json",
         "package": "shx echo '{ \"type\": \"commonjs\" }' > lib/cjs/package.json",
-        "fmt": "prettier --write '{*,**/*}.{ts,tsx,js,jsx,json}'",
+        "fmt": "prettier --write '{*,**/*}.{ts,tsx,js,jsx,json}' --cache",
         "lint": "prettier --check '{*,**/*}.{ts,tsx,js,jsx,json}' && eslint . --cache",
         "lint:fix": "yarn run fmt && eslint --fix . --cache"
     },