--- a/packages/wallet-standard/package.json
+++ b/packages/wallet-standard/package.json
@@ -35,7 +35,6 @@
         "@wallet-standard/features": "^1.0.0",
         "eslint": "8.22.0",
         "eslint-config-prettier": "^8.5.0",
-        "eslint-plugin-prettier": "^4.2.1",
         "eslint-plugin-require-extensions": "^0.1.1",
         "prettier": "^2.7.1",
         "shx": "^0.3.4",