eslint was very slow on save. for vscode folks, slow is the norm. for vim/emacs folks, slow is not the norm. this isn't a silver bullet, but it's an improvement

## the process 
running `TIMING=1 yarn lint` resulted in discovering that `prettier/prettier` takes up a whopping second on every save (or every time eslint runs, so all the time).

`eslint-plugin-prettier` will run prettier on the underlying code first, causing the slow-down. By moving towards `eslint-config-prettier` we're able to prevent that from happening and only check prettier styles.

## the result

BEFORE:

```
@coral-xyz/wallet-standard:lint: Rule                                            | Time (ms) | Relative
@coral-xyz/wallet-standard:lint: :-----------------------------------------------|----------:|--------:
@coral-xyz/wallet-standard:lint: prettier/prettier                               |  1098.920 |    97.9%
@coral-xyz/wallet-standard:lint: prefer-spread                                   |     4.016 |     0.4%
@coral-xyz/wallet-standard:lint: @typescript-eslint/no-unused-vars               |     3.514 |     0.3%
@coral-xyz/wallet-standard:lint: @typescript-eslint/no-empty-function            |     1.616 |     0.1%
@coral-xyz/wallet-standard:lint: no-misleading-character-class                   |     1.085 |     0.1%
@coral-xyz/wallet-standard:lint: no-control-regex                                |     0.908 |     0.1%
@coral-xyz/wallet-standard:lint: @typescript-eslint/consistent-type-imports      |     0.875 |     0.1%
@coral-xyz/wallet-standard:lint: no-debugger                                     |     0.868 |     0.1%
@coral-xyz/wallet-standard:lint: @typescript-eslint/no-non-null-assertion        |     0.602 |     0.1%
@coral-xyz/wallet-standard:lint: @typescript-eslint/adjacent-overload-signatures |     0.588 |     0.1%
```

AFTER:

```
All matched files use Prettier code style!
Rule                                       | Time (ms) | Relative
:------------------------------------------|----------:|--------:
@typescript-eslint/no-inferrable-types     |     4.857 |    24.8%
@typescript-eslint/no-unused-vars          |     3.092 |    15.8%
@typescript-eslint/no-empty-function       |     1.213 |     6.2%
no-misleading-character-class              |     0.989 |     5.0%
no-control-regex                           |     0.771 |     3.9%
@typescript-eslint/consistent-type-imports |     0.619 |     3.2%
no-useless-backreference                   |     0.546 |     2.8%
@typescript-eslint/no-non-null-assertion   |     0.526 |     2.7%
@typescript-eslint/no-loss-of-precision    |     0.453 |     2.3%
@typescript-eslint/triple-slash-reference  |     0.416 |     2.1%
```

## summary 
- adds `--cache` to `eslint`
- adds `--cache` to `prettier` (introduced in 2.7)
- replaces `eslint-plugin-prettier` with `eslint-config-prettier` to speed up eslint
- fixes `common-public` eslint parsing error (noticed this while running the command) by adding `parserOptions`
- adds `.eslintcache` to `.gitignore`

## possible next steps
- run `eslint --fix` on everything and reduce the amount of warnings that show up. duplicate imports/import-order is the next time-consuming process and that seems like an easy win
- adds paths / references to tsconfig.json and fix the next elephant in the room :eyes: